flaskbp/requirements.txt
Flask==1.1.1
geoip2==4.1.0
pymongo==3.11.2
pyecharts==1.9.0
dnspython==2.1.0
